+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -38,6 +38,7 @@ export * from './FloatingActionButton';
 export * from './PullToRefreshBox';
 export * from './RadioButton';
 export * from './Surface';
+export * from './SwipeToDismissBox';
 export { type TextProps, Text } from './Text';
 export * from './Tooltip';